Fixes for Kafka 3.0
