+++ .env
@@ -1,1 +1,1 @@
-KAFKA_CONFLUENT_VERSION=6.1.0
+KAFKA_CONFLUENT_VERSION=7.0.0
