+++ Dockerfile
@@ -18,7 +18,6 @@ WORKDIR /code
 ENV GRADLE_OPTS="-Dorg.gradle.daemon=false"
 
 COPY ./build.gradle.kts ./gradle.properties ./settings.gradle.kts /code/
-COPY ./deprecated-javax/build.gradle.kts /code/deprecated-javax/
 
 RUN gradle downloadDockerDependencies --no-watch-fs
