+++ gateway.yml
@@ -18,13 +18,14 @@ kafka:
     # Kafka producer settings. Read from https://kafka.apache.org/documentation/#producerconfigs.
     producer:
         bootstrap.servers: kafka-1:9092
-    # Kafka Admin Client settings. Read from https://kafka.apache.org/documentation/#adminclientconfigs.
+        security.protocol: PLAINTEXT
+        # Kafka Admin Client settings. Read from https://kafka.apache.org/documentation/#adminclientconfigs.
     #admin:
         # bootstrap server property is copied from the producer settings if none is provided.
         #bootstrap.servers: kafka-1:9092
     # Kafka serialization settings, used in KafkaAvroSerializer. Read from [io.confluent.kafka.serializers.AbstractKafkaSchemaSerDeConfig].
     serialization:
-        schema.registry.url: http://schema-registry-1:8081
+        schema.registry.url: http://schema-registry:8081
 
 # Authorization settings
 auth:
