+++ gradle.properties
@@ -1,6 +1,6 @@
 org.gradle.jvmargs=-Xmx3072m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
 kotlin.code.style=official
-dockerComposeStopContainers=true
+dockerComposeStopContainers=false
 
 kotlinVersion=1.6.0
 okhttp3Version=4.9.3
@@ -11,9 +11,11 @@ jacksonVersion=2.12.5
 slf4jVersion=1.7.32
 log4j2Version=2.15.0
 lzfseVersion=0.1.1
+radarAuthVersion=0.8.0
 avroVersion=1.11.0
 kafkaVersion=3.0.0
 confluentVersion=7.0.1
+
+mockitoKotlinVersion=4.0.0
 junitVersion=5.8.2
-radarAuthVersion=0.8.0
 grizzlyVersion=3.0.1
