+++ settings.gradle.kts
@@ -1,7 +1,5 @@
 rootProject.name = "radar-gateway"
 
-include(":deprecated-javax")
-
 pluginManagement {
     val kotlinVersion: String by settings
     plugins {
