+++ src/integrationTest/docker/.env
@@ -1,2 +1,2 @@
-KAFKA_CONFLUENT_VERSION=6.1.0
+KAFKA_CONFLUENT_VERSION=7.0.0
 RADAR_GATEWAY_TAG=SNAPSHOT
