+++ src/main/kotlin/org/radarbase/gateway/inject/ManagementPortalEnhancerFactory.kt
@@ -1,6 +1,6 @@
 package org.radarbase.gateway.inject
 
-import org.radarbase.gateway.Config
+import org.radarbase.gateway.config.GatewayConfig
 import org.radarbase.jersey.auth.AuthConfig
 import org.radarbase.jersey.auth.MPConfig
 import org.radarbase.jersey.enhancer.EnhancerFactory
@@ -8,7 +8,7 @@ import org.radarbase.jersey.enhancer.JerseyResourceEnhancer
 import org.radarbase.jersey.enhancer.Enhancers
 
 /** This binder needs to register all non-Jersey classes, otherwise initialization fails. */
-class ManagementPortalEnhancerFactory(private val config: Config) : EnhancerFactory {
+class ManagementPortalEnhancerFactory(private val config: GatewayConfig) : EnhancerFactory {
     override fun createEnhancers(): List<JerseyResourceEnhancer> {
         val authConfig = AuthConfig(
             managementPortal = MPConfig(
