+++ src/main/kotlin/org/radarbase/gateway/io/SizeLimitInterceptor.kt
@@ -7,7 +7,7 @@ import jakarta.ws.rs.core.Context
 import jakarta.ws.rs.ext.Provider
 import jakarta.ws.rs.ext.ReaderInterceptor
 import jakarta.ws.rs.ext.ReaderInterceptorContext
-import org.radarbase.gateway.Config
+import org.radarbase.gateway.config.GatewayConfig
 import org.radarbase.gateway.inject.ProcessAvro
 
 /**
@@ -17,7 +17,7 @@ import org.radarbase.gateway.inject.ProcessAvro
 @ProcessAvro
 @Singleton
 @Priority(Priorities.ENTITY_CODER + 100)
-class SizeLimitInterceptor(@Context private val config: Config) : ReaderInterceptor {
+class SizeLimitInterceptor(@Context private val config: GatewayConfig) : ReaderInterceptor {
     override fun aroundReadFrom(context: ReaderInterceptorContext): Any {
         context.inputStream = LimitedInputStream(context.inputStream, config.server.maxRequestSize)
         return context.proceed()
