+++ src/main/kotlin/org/radarbase/gateway/kafka/ProducerPoolFactory.kt
@@ -2,10 +2,10 @@ package org.radarbase.gateway.kafka
 
 import jakarta.ws.rs.core.Context
 import org.glassfish.jersey.internal.inject.DisposableSupplier
-import org.radarbase.gateway.Config
+import org.radarbase.gateway.config.GatewayConfig
 
 class ProducerPoolFactory(
-    @Context private val config: Config,
+    @Context private val config: GatewayConfig,
 ) : DisposableSupplier<ProducerPool> {
     override fun get(): ProducerPool = ProducerPool(config)
